requirements.txt
gitpython==3.1.46
pyside6==6.10.2
jinja2==3.1.6
pyopengl==3.1.10
requests==2.32.5